requirements.txt
alright
colorama
gspread
phonenumbers
polars
PyYAML
selenium
webdriver-manager
